requirements.txt
Django>=3.1,<3.2
wagtail>=2.12,<2.13

django-allauth
django-extensions
djangorestframework

